Move command generation to a command.
