+++ tests/TestCase/Command/SimpleBakeCommandTest.php
@@ -16,6 +16,7 @@ declare(strict_types=1);
 namespace Bake\Test\TestCase\Command;
 
 use Bake\Test\TestCase\TestCase;
+use Cake\Console\Command;
 use Cake\Core\Plugin;
 
 /**
